Restore service file instead of folder
